Add multi-step booking process with progress bar and form components
